remove  recruitment page, create pricing container
