--- app/containers/CompanyRecruitment/constants.js
@@ -1,11 +1,0 @@
-export const CHANGE_INPUT = 'app/CompanyRecruitment/CHANGE_INPUT';
-
-export const CHANGE_STEP = 'app/CompanyRecruitment/CHANGE_STEP';
-
-export const INPUT_ERROR = 'app/CompanyRecruitment/INPUT_ERROR';
-
-export const RESET_FORM = 'app/CompanyRecruitment/RESET_FORM';
-
-export const SEND_FORM = 'app/CompanyRecruitment/SEND_FORM';
-export const SEND_FORM_FAILURE = 'app/CompanyRecruitment/SEND_FORM_FAILURE';
-export const SEND_FORM_SUCCESS = 'app/CompanyRecruitment/SEND_FORM_SUCCESS';
