+++ app/containers/Pricing/styledComponents.js
@@ -1,8 +1,6 @@
 import styled from 'styled-components';
 
-import { blueColor } from 'defaultStyleHelper';
-
 export const ViewContainer = styled.div`
-  background: ${blueColor};
+  display: flex;
   width: 100%;
 `;
